dbd/requirements.txt
altair==5.2.0
APScheduler==3.10.4
apturl==0.5.2
attrs==23.2.0
bcrypt==4.1.2
blinker==1.7.0
boto3==1.28.71
botocore==1.31.85
Brlapi==0.8.3
cachelib==0.12.0
cachetools==5.3.2
certifi==2024.2.2
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
duplicity==0.8.21
fasteners==0.14.1
Flask==3.0.2
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-Session==0.7.0
Flask-SQLAlchemy==3.1.1
future==0.18.2
gitdb==4.0.11
GitPython==3.1.41
greenlet==3.0.3
gyp==0.1
httplib2==0.20.2
idna==3.6
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.3
jmespath==1.0.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
monotonic==1.6
more-itertools==8.10.0
msgspec==0.18.6
netifaces==0.11.0
numpy==1.26.3
oauthlib==3.2.2
olefile==0.46
packaging==23.2
pandas==2.2.0
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
playsound==1.3.0
protobuf==4.25.2
ptyprocess==0.7.0
pyarrow==15.0.0
pycairo==1.20.1
pycups==2.0.1
pydeck==0.8.1b0
Pygments==2.17.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-vlc==3.0.20123
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
referencing==0.33.0
reportlab==3.6.8
requests==2.31.0
rich==13.7.0
rpds-py==0.17.1
s3transfer==0.7.0
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.28
streamlit==1.27.2
systemd-python==234
tenacity==8.2.3
toml==0.10.2
toolz==0.12.1
tornado==6.4
typing_extensions==4.10.0
tzdata==2023.4
tzlocal==5.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.2.1
usb-creator==0.3.7
validators==0.22.0
wadllib==1.3.6
watchdog==3.0.0
Werkzeug==3.0.1
xdg==5
xkit==0.0.0
zipp==1.0.0
